Move AgentSubscriber and ILLMService interfaces to more appropriate files
