+++ src/ai/llm/events/event-manager.ts
@@ -1,4 +1,5 @@
-import { ILLMService, AgentSubscriber } from '../types.js';
+import { AgentSubscriber } from './types.js';
+import { ILLMService } from '../services/types.js';
 import { logger } from '../../../utils/logger.js';
 /**
  * Manages event subscriptions and distribution for the LLM service
