+++ src/ai/llm/services/factory.ts
@@ -1,5 +1,5 @@
 import { ClientManager } from '../../../client/manager.js';
-import { ILLMService } from '../types.js';
+import { ILLMService } from './types.js';
 import { LLMConfig } from '../../../config/types.js';
 import { logger } from '../../../utils/logger.js';
 import { openai } from '@ai-sdk/openai';
